+++ priv/repo/migrations/20230315072923_add_file_info_csv_bulk_update_events.exs
@@ -1,0 +1,9 @@
+defmodule TdDd.Repo.Migrations.AddFileInfoCsvBulkUpdateEvents do
+  use Ecto.Migration
+
+  def change do
+    alter table("csv_bulk_update_events") do
+      add(:filename, :string, null: true)
+    end
+  end
+end