+++ test/td_dd_web/controllers/data_structure_controller_test.exs
@@ -5,6 +5,7 @@ defmodule TdDdWeb.DataStructureControllerTest do
   import Mox
   import Routes
 
+  alias Path
   alias TdDd.DataStructures
   alias TdDd.DataStructures.DataStructure
   alias TdDd.DataStructures.StructureNotes
@@ -14,6 +15,13 @@ defmodule TdDdWeb.DataStructureControllerTest do
   @template_with_multifields_name "data_structure_controller_test_template_with_multifields"
   @receive_timeout 500
   @protected DataStructures.protected()
+
+  defp plug_upload(path) do
+    %Plug.Upload{
+      path: path,
+      filename: Path.basename(path)
+    }
+  end
 
   # function that returns the function to be injected
   def notify_callback do
@@ -715,7 +723,7 @@ defmodule TdDdWeb.DataStructureControllerTest do
 
       conn
       |> post(data_structure_path(conn, :bulk_upload_domains),
-        structures_domains: %Plug.Upload{path: "test/fixtures/td4535/structures_domains_good.csv"}
+        structures_domains: plug_upload("test/fixtures/td4535/structures_domains_good.csv")
       )
       |> json_response(:forbidden)
     end
@@ -736,9 +744,7 @@ defmodule TdDdWeb.DataStructureControllerTest do
       data =
         conn
         |> post(data_structure_path(conn, :bulk_upload_domains),
-          structures_domains: %Plug.Upload{
-            path: "test/fixtures/td4535/structures_domains_good.csv"
-          }
+          structures_domains: plug_upload("test/fixtures/td4535/structures_domains_good.csv")
         )
         |> json_response(:ok)
 
@@ -792,9 +798,8 @@ defmodule TdDdWeb.DataStructureControllerTest do
       data =
         conn
         |> post(data_structure_path(conn, :bulk_upload_domains),
-          structures_domains: %Plug.Upload{
-            path: "test/fixtures/td4535/structures_domains_permissions.csv"
-          }
+          structures_domains:
+            plug_upload("test/fixtures/td4535/structures_domains_permissions.csv")
         )
         |> json_response(:ok)
 
@@ -831,9 +836,8 @@ defmodule TdDdWeb.DataStructureControllerTest do
       data =
         conn
         |> post(data_structure_path(conn, :bulk_upload_domains),
-          structures_domains: %Plug.Upload{
-            path: "test/fixtures/td4535/structures_domains_bad_header.csv"
-          }
+          structures_domains:
+            plug_upload("test/fixtures/td4535/structures_domains_bad_header.csv")
         )
         |> json_response(:unprocessable_entity)
 
@@ -860,9 +864,8 @@ defmodule TdDdWeb.DataStructureControllerTest do
       data =
         conn
         |> post(data_structure_path(conn, :bulk_upload_domains),
-          structures_domains: %Plug.Upload{
-            path: "test/fixtures/td4535/structures_domains_bad_missing_external_id.csv"
-          }
+          structures_domains:
+            plug_upload("test/fixtures/td4535/structures_domains_bad_missing_external_id.csv")
         )
         |> json_response(:ok)
 
@@ -918,9 +921,8 @@ defmodule TdDdWeb.DataStructureControllerTest do
       data =
         conn
         |> post(data_structure_path(conn, :bulk_upload_domains),
-          structures_domains: %Plug.Upload{
-            path: "test/fixtures/td4535/structures_domains_warning_inexistent.csv"
-          }
+          structures_domains:
+            plug_upload("test/fixtures/td4535/structures_domains_warning_inexistent.csv")
         )
         |> json_response(:ok)
 
@@ -1070,7 +1072,7 @@ defmodule TdDdWeb.DataStructureControllerTest do
 
       conn
       |> post(data_structure_path(conn, :bulk_update_template_content),
-        structures: %Plug.Upload{path: "test/fixtures/td3787/upload_unprocessable_entity.csv"}
+        structures: plug_upload("test/fixtures/td3787/upload_unprocessable_entity.csv")
       )
       |> json_response(:unprocessable_entity)
     end
@@ -1089,7 +1091,7 @@ defmodule TdDdWeb.DataStructureControllerTest do
              } =
                conn
                |> post(data_structure_path(conn, :bulk_update_template_content),
-                 structures: %Plug.Upload{path: "test/fixtures/td4100/upload.csv"}
+                 structures: plug_upload("test/fixtures/td4100/upload.csv")
                )
                |> json_response(:accepted)
 
@@ -1124,9 +1126,7 @@ defmodule TdDdWeb.DataStructureControllerTest do
              } =
                conn
                |> post(data_structure_path(conn, :bulk_update_template_content),
-                 structures: %Plug.Upload{
-                   path: "test/fixtures/td4100/upload_with_one_warning.csv"
-                 }
+                 structures: plug_upload("test/fixtures/td4100/upload_with_one_warning.csv")
                )
                |> json_response(:accepted)
 
@@ -1171,9 +1171,7 @@ defmodule TdDdWeb.DataStructureControllerTest do
              } =
                conn
                |> post(data_structure_path(conn, :bulk_update_template_content),
-                 structures: %Plug.Upload{
-                   path: "test/fixtures/td4100/upload_with_multiple_warnings.csv"
-                 }
+                 structures: plug_upload("test/fixtures/td4100/upload_with_multiple_warnings.csv")
                )
                |> json_response(:accepted)
 
@@ -1223,9 +1221,8 @@ defmodule TdDdWeb.DataStructureControllerTest do
              } =
                conn
                |> post(data_structure_path(conn, :bulk_update_template_content),
-                 structures: %Plug.Upload{
-                   path: "test/fixtures/td4100/upload_with_invalid_external_id.csv"
-                 }
+                 structures:
+                   plug_upload("test/fixtures/td4100/upload_with_invalid_external_id.csv")
                )
                |> json_response(:accepted)
 
@@ -1257,7 +1254,7 @@ defmodule TdDdWeb.DataStructureControllerTest do
 
       conn
       |> post(data_structure_path(conn, :bulk_update_template_content),
-        structures: %Plug.Upload{path: "test/fixtures/td3787/upload.csv"}
+        structures: plug_upload("test/fixtures/td3787/upload.csv")
       )
       |> response(:accepted)
 
@@ -1266,6 +1263,7 @@ defmodule TdDdWeb.DataStructureControllerTest do
       assert [
                %{
                  "csv_hash" => _csv_hash,
+                 "filename" => "upload.csv",
                  "status" => "COMPLETED",
                  "task_reference" => _task_reference,
                  "response" => %{"ids" => _, "errors" => []}
@@ -1293,7 +1291,7 @@ defmodule TdDdWeb.DataStructureControllerTest do
 
       conn
       |> post(data_structure_path(conn, :bulk_update_template_content),
-        structures: %Plug.Upload{path: "test/fixtures/td4548/upload_with_multifields.csv"}
+        structures: plug_upload("test/fixtures/td4548/upload_with_multifields.csv")
       )
       |> response(:accepted)
 
@@ -1330,7 +1328,7 @@ defmodule TdDdWeb.DataStructureControllerTest do
 
       conn
       |> post(data_structure_path(conn, :bulk_update_template_content),
-        structures: %Plug.Upload{path: "test/fixtures/td3787/upload.csv"}
+        structures: plug_upload("test/fixtures/td3787/upload.csv")
       )
       |> response(:forbidden)
     end
@@ -1353,7 +1351,7 @@ defmodule TdDdWeb.DataStructureControllerTest do
 
       conn
       |> post(data_structure_path(conn, :bulk_update_template_content),
-        structures: %Plug.Upload{path: "test/fixtures/td3787/upload.csv"}
+        structures: plug_upload("test/fixtures/td3787/upload.csv")
       )
       |> response(:accepted)
 
@@ -1398,7 +1396,7 @@ defmodule TdDdWeb.DataStructureControllerTest do
 
       conn
       |> post(data_structure_path(conn, :bulk_update_template_content),
-        structures: %Plug.Upload{path: "test/fixtures/td3787/upload.csv"}
+        structures: plug_upload("test/fixtures/td3787/upload.csv")
       )
       |> response(:accepted)
 
@@ -1445,7 +1443,7 @@ defmodule TdDdWeb.DataStructureControllerTest do
 
       conn
       |> post(data_structure_path(conn, :bulk_update_template_content),
-        structures: %Plug.Upload{path: "test/fixtures/td3787/upload.csv"}
+        structures: plug_upload("test/fixtures/td3787/upload.csv")
       )
       |> response(:accepted)
 
@@ -1484,7 +1482,7 @@ defmodule TdDdWeb.DataStructureControllerTest do
 
       conn
       |> post(data_structure_path(conn, :bulk_update_template_content),
-        structures: %Plug.Upload{path: "test/fixtures/td3787/upload.csv"},
+        structures: plug_upload("test/fixtures/td3787/upload.csv"),
         auto_publish: "true"
       )
       |> response(:accepted)
@@ -1525,7 +1523,7 @@ defmodule TdDdWeb.DataStructureControllerTest do
 
       conn
       |> post(data_structure_path(conn, :bulk_update_template_content),
-        structures: %Plug.Upload{path: "test/fixtures/td3787/upload.csv"},
+        structures: plug_upload("test/fixtures/td3787/upload.csv"),
         auto_publish: "true"
       )
       |> response(:forbidden)